TASK: Finalize fix tasks

+++ src/EditorconfigChecker/Fix/FinalNewlineFix.php
@@ -8,11 +8,12 @@ class FinalNewlineFix
      * Insert a final newline at the end of the file
      *
      * @param string $filename
+     * @param string $eolChar
      * @return boolean
      */
     public static function insert($filename, $eolChar)
     {
-        if (is_file($filename)) {
+        if (is_file($filename) && $eolChar) {
             return file_put_contents($filename, $eolChar, FILE_APPEND);
         }
